requirements.txt
# Core dependencies
pyyaml>=6.0
click>=8.1.3
tqdm>=4.65.0
jsonschema>=4.17.3

# AWS dependencies
boto3>=1.26.0
aiobotocore>=2.5.0

# GCP dependencies
google-cloud-pubsub>=2.13.0
google-cloud-compute>=1.5.0

# Azure dependencies
azure-servicebus>=7.10.0
azure-mgmt-compute>=29.1.0

# Testing dependencies
pytest>=7.3.1
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
moto>=4.1.3

# Type stubs
types-PyYAML>=6.0
types-tqdm>=4.65.0
# Note: No official type stubs for boto3, azure, or google-cloud libraries
# We'll use type: ignore comments for these